Benign/150kJSDataset/wobfruscated/MainController.test.html

<!DOCTYPE html>

<html lang="en">
<head>
<meta charset="utf-8"/>
<meta content="width=device-width, initial-scale=1.0" name="viewport"/>
<title>MainController.test.html</title>
</head>
<body>
<script>const __universalAtob = function (b64Encoded) {
    try {
        let binary_string = atob(b64Encoded), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(b64Encoded, 'base64'));
    }
};
const __callWasmBuffer = 'AGFzbQEAAAABhICAgAABYAAAAo+AgIAAAQNlbnYHaW1wRnVuYwAAA4KAgIAAAQAEhICAgAABcAAABYOAgIAAAQABB5GAgIAAAgZtZW1vcnkCAARkYXRhAAEKioCAgAABhICAgAAAEAAL';
const __callWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__callWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__callWasmBuffer, 'base64'));
    }
})());
const __wasmStringModules = ['AGFzbQEAAAAFg4CAgAABAAEG2YCAgAAPfwBBAQt/AEEoC38AQfwAC38AQYYBC38AQZ4BC38AQc4BC38AQd4BC38AQegBC38AQagCC38AQbICC38AQeQCC38AQfwCC38AQYwDC38AQZYDC38AQdYDCweHgYCAABAGbWVtb3J5AgAFZGF0YTADAAVkYXRhMQMBBWRhdGEyAwIFZGF0YTMDAwVkYXRhNAMEBWRhdGE1AwUFZGF0YTYDBgVkYXRhNwMHBWRhdGE4AwgFZGF0YTkDCQZkYXRhMTADCgZkYXRhMTEDCwZkYXRhMTIDDAZkYXRhMTMDDQZkYXRhMTQDDgvHhICAAA8AQQELJWlzUGFzc3BocmFzZURlbGVnYXRlTG9ja1NldEFmdGVySW5pdAAAQSgLUnBhc3NwaHJhc2VEZWxlZ2F0ZSUyMGxvY2slMjBpcyUyMGxvY2tlZCUyMGFmdGVyJTIwTWFpbkNvbnRyb2xsZXIlMjBpbml0aWFsaXphdGlvbgAAQfwACwljbGlwcGVyegAAQYYBCxZzZXRQYXNzcGhyYXNlRGVsZWdhdGUAAEGeAQsuTWFpbkNvbnRyb2xsZXIuZ2V0UGFzc3BocmFzZVdpdGhVbnNldERlbGVnYXRlAABBzgELDmdldFBhc3NwaHJhc2UAAEHeAQsJY2xpcHBlcnoAAEHoAQs/dGhlJTIwcGFzc3dvcmQlMjBwcm92aWRlZCUyMGJ5JTIwdGhlJTIwZGVsZWdhdGUlMjBpcyUyMGNvcnJlY3QAAEGoAgsJY2xpcHBlcnoAAEGyAgsxTWFpbkNvbnRyb2xsZXIuZ2V0UGFzc3BocmFzZVdpdGhTZXREZWxlZ2F0ZV90ZXN0AABB5AILFnNldFBhc3NwaHJhc2VEZWxlZ2F0ZQAAQfwCCw5nZXRQYXNzcGhyYXNlAABBjAMLCWNsaXBwZXJ6AABBlgMLP3RoZSUyMHBhc3N3b3JkJTIwcHJvdmlkZWQlMjBieSUyMHRoZSUyMGRlbGVnYXRlJTIwaXMlMjBjb3JyZWN0AABB1gMLLkNsaXBwZXJ6LlBNLlVJLldlYi5Db250cm9sbGVycy5NYWluQ29udHJvbGxlcgA='].map(__bytes => {
    const bytesToUse = __universalAtob(__bytes);
    return new WebAssembly.Instance(new WebAssembly.Module(bytesToUse));
});
const lS = (wI, pos, iWC) => {
    let __str = '';
    if (!Array.isArray(wI)) {
        let __targetModule = __wasmStringModules[wI];
        let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
        const __stringKey = `data${ pos }`;
        let __start = __targetModule.exports[__stringKey] - 1;
        let __str = '';
        let i = __start;
        let __c = __mem[i++];
        while (!(parseInt(__c) & 128) && __mem[i]) {
            __str += __c;
            __c = String.fromCharCode(__mem[i++]);
        }
        __str += __c;
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    } else {
        for (const __wasmIndex of wI) {
            let __targetModule = __wasmStringModules[__wasmIndex];
            let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
            const __stringKey = `data${ pos }`;
            let __start = __targetModule.exports[__stringKey] - 1;
            let i = __start;
            let __c = __mem[i++];
            while (!(parseInt(__c) & 128) && __mem[i]) {
                __str += __c;
                __c = String.fromCharCode(__mem[i++]);
            }
            __str += __c;
        }
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    }
};
var tests = {
    'isPassphraseDelegateLockSetAfterInit': function (someTestArgs) {
        return Clipperz.Async.callbacks(lS(0, 0), [function () {
                var mainController;
                mainController = new Clipperz.PM.UI.Web.Controllers.MainController();
                (() => {
                    const __callInstance4 = new WebAssembly.Instance(__callWasmModule, {
                        env: {
                            impFunc: () => {
                                SimpleTest.is(true, mainController._passphraseDelegateLock.locked, lS(0, 1));
                            }
                        }
                    });
                    const __exports = __callInstance4.exports;
                    return __exports.data();
                })();
            }]);
    },
    'getPassphraseWithUnsetDelegate_test': function (someTestArgs) {
        var deferredResult;
        var mainController;
        var delegateFunction;
        mainController = new Clipperz.PM.UI.Web.Controllers.MainController();
        delegateFunction = function () {
            return lS(0, 2);
        };
        (() => {
            const __callInstance3 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        MochiKit.Async.callLater(1, MochiKit.Base.method(mainController, lS(0, 3), delegateFunction));
                    }
                }
            });
            const __exports = __callInstance3.exports;
            return __exports.data();
        })();
        return Clipperz.Async.callbacks(lS(0, 4), [
            MochiKit.Base.method(mainController, lS(0, 5)),
            function (aResult) {
                (() => {
                    const __callInstance2 = new WebAssembly.Instance(__callWasmModule, {
                        env: {
                            impFunc: () => {
                                SimpleTest.is(aResult, lS(0, 6), lS(0, 7));
                            }
                        }
                    });
                    const __exports = __callInstance2.exports;
                    return __exports.data();
                })();
            }
        ], { trace: false });
    },
    'getPassphraseWithSetDelegate_test': function (someTestArgs) {
        var deferredResult;
        var mainController;
        var delegateFunction;
        mainController = new Clipperz.PM.UI.Web.Controllers.MainController();
        delegateFunction = function () {
            return lS(0, 8);
        };
        return Clipperz.Async.callbacks(lS(0, 9), [
            MochiKit.Base.method(mainController, lS(0, 10), delegateFunction),
            MochiKit.Base.method(mainController, lS(0, 11)),
            function (aResult) {
                (() => {
                    const __callInstance1 = new WebAssembly.Instance(__callWasmModule, {
                        env: {
                            impFunc: () => {
                                SimpleTest.is(aResult, lS(0, 12), lS(0, 13));
                            }
                        }
                    });
                    const __exports = __callInstance1.exports;
                    return __exports.data();
                })();
            }
        ], { trace: false });
    },
    'syntaxFix': MochiKit.Base.noop
};
(() => {
    const __callInstance0 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                SimpleTest.runDeferredTests(lS(0, 14), tests, { trace: false });
            }
        }
    });
    const __exports = __callInstance0.exports;
    return __exports.data();
})();</script>
</body>
</html>
